+++ src/panels/config/ha-panel-config.ts
@@ -482,7 +482,7 @@ class HaPanelConfig extends HassRouterPage {
         tag: "matter-config-panel",
         load: () =>
           import(
-            "./integrations/integration-panels/matter/matter-config-panel"
+            "./integrations/integration-panels/matter/matter-config-dashboard"
           ),
       },
       thread: {
